feat : 공통코드 조건 조회 JPA 추가

+++ src/main/java/com/pintalk/common/entity/ComCode.java
@@ -14,7 +14,7 @@ public class ComCode extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY, generator = "COMCODE_GENERATOR")
-    private Integer no;
+    private Long no;
 
     private String codeValue;
     private String codeGroup;
